Add comment prefix handling functions
Introduce `commentPrefixPattern` and `calculateCommentPrefixRemovalEnd`
to manage leading semicolons in comments. This improves the ability to
toggle comments by accurately determining where to remove comment
prefixes, enhancing the overall functionality of the comment system.
